feat: sam package `click` wiring
- added unit tests for the click wiring
